requirements.txt
kiteconnect>=4.1.0
requests>=2.28.0
pandas>=1.5.0
numpy>=1.23.0
python-dotenv>=1.0.0
schedule>=1.2.0
pyotp>=2.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
